requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
xgboost
nltk
matplotlib
plotly
wordcloud
seaborn
tk
scipy
